3D metrics should be ok + PSNR mismatch fixed

+++ iMeritsADMMLim.py
@@ -28,7 +28,7 @@ class iMeritsADMMLim(vGeneral):
         self.phantom_ROI = self.get_phantom_ROI()
 
         #Loading Ground Truth image to compute metrics
-        self.image_gt = self.fijii_np(self.subroot_data + 'Data/database_v2/' + self.phantom + '/' + self.phantom + '.raw',shape=(self.PETImage_shape),type='<f').astype(np.float64)
+        self.image_gt = self.fijii_np(self.subroot_data + 'Data/database_v2/' + self.phantom + '/' + self.phantom + '.raw',shape=(self.PETImage_shape),type='<f')
         if settings_config["FLTNB"] == "double":
             self.image_gt.astype(np.float64)
